+++ src/OneBlinkFormBase.tsx
@@ -637,6 +637,7 @@ function OneBlinkFormBase({
                                       onChangeElements={handleChangeElements}
                                       onChangeModel={handleChangeModel}
                                       formElementsCtrl={rootFormElementsCtrl}
+                                      idPrefix=""
                                     />
                                   </div>
                                 ),