+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.3] - 2021-04-15
+
 ### Changed
 
 - Tooltip library to use material-ui tooltip
@@ -18,6 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - inputs in sub forms in repeatable sets having duplicate `id` properties, causing a bug with `date`, `datetime`, `time` elements not setting the value selected on the submission data
+
+### Dependencies
+
+- no longer depend upon [react-tooltip](https://www.npmjs.com/package/react-tooltip)
+
+- depend upon [@material-ui/core](https://www.npmjs.com/package/@material-ui/core) [4.11.3](https://github.com/mui-org/material-ui/releases/tag/v4.11.3)
 
 ## [0.2.2] - 2021-03-31